Use Hellvetica font
Hellvetica[1] is Helvetica but with bad kerning.

[1]: https://hellveticafont.com/

It is only provided in TTF format, so that'll have to do. Tested with
latest Firefox and Chrome on Windows, and Firefox on Android.

Put the CSS in the head tag instead of templates/css/ so that we can
gate it on the DNF header.